Crate os_info

os_info

Provides interfaces for getting information about the current operating system, such as type, version and edition.

Structs

Info
Holds information about operating system (type, version, etc.).

Version
Operating system version including version number and optional edition.

Enums

Bitness
Operating system architecture in terms of how many bits compose the basic values it can deal with.

Type
A list of supported operating system types.

VersionType
Operating system version.

Functions

get
Returns information about the current operating system (type, version, edition, etc.).